+++ cct-system/src/main/java/com/cct/model/Usuario.java
@@ -4,11 +4,6 @@ import java.io.Serializable;
 import javax.persistence.*;
 import java.util.List;
 
-
-/**
- * The persistent class for the usuario database table.
- * 
- */
 @Entity
 @NamedQuery(name="Usuario.findAll", query="SELECT u FROM Usuario u")
 public class Usuario implements Serializable {
@@ -16,7 +11,7 @@ public class Usuario implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private Integer idusuario;
+	private Long idusuario;
 
 	private String email;
 
@@ -29,11 +24,11 @@ public class Usuario implements Serializable {
 	public Usuario() {
 	}
 
-	public Integer getIdusuario() {
+	public Long getIdusuario() {
 		return this.idusuario;
 	}
 
-	public void setIdusuario(Integer idusuario) {
+	public void setIdusuario(Long idusuario) {
 		this.idusuario = idusuario;
 	}
